6: jsonify error messages

+++ api/v1/views/states.py
@@ -28,21 +28,18 @@ def get_state(state_id):
                  strict_slashes=False)
 def delete_state(state_id):
     """Delete a State"""
-    try:
-        state = storage.get(State, state_id)
+    state = storage.get(State, state_id)
+    if state:
         storage.delete(state)
         storage.save()
         return jsonify({}), 200
-    except Exception as e:
+    else:
         abort(404)
 
 
 @app_views.route('/states', methods=['POST'], strict_slashes=False)
 def post_state():
     """Creates a State object"""
-    ## get the dict from the request {name}
-    ## check if dict is valid
-    ## return new state with 201
     req = request.get_json()
     if req:
         if "name" in req.keys():
@@ -51,25 +48,25 @@ def post_state():
             state.save()
             return jsonify(state.to_dict()), 201
         else:
-            return "Missing name", 400
+            return jsonify({"error": "Missing name"}), 400
     else:
-        return "Not a JSON", 400
+        return jsonify({"error": "Not a JSON"}), 400
 
 
 @app_views.route('/states/<state_id>', methods=['PUT'], strict_slashes=False)
 def put_state(state_id):
     """Updates a State object"""
-    try:
-        state = storage.get(State, state_id)
-    except Exception as e:
-        abort(404)
     req = request.get_json()
     if req:
-        state.name = req["name"]
-        state.save()
-        return jsonify(state.to_dict()), 200
+        try:
+            state = storage.get(State, state_id)
+            state.name = req["name"]
+            state.save()
+            return jsonify(state.to_dict()), 200
+        except Exception as e:
+            abort(404)
     else:
-        return "Not a JSON", 400
+        return jsonify({"error": "Not a JSON"}), 400
 
 
 if __name__ == '__main__':
